Added eureka and actuator , swagger

+++ EmployeeManagement/src/main/java/com/emsb/service/EmployeesService.java
@@ -12,62 +12,71 @@ import com.emsb.repository.EmployeesRepository;
 
 @Service
 public class EmployeesService {
-	
+
 	@Autowired
 	private EmployeesRepository employeesrepo;
-	
-	/*Adding a employees*/
-	public Employees addEmployee(Employees employee) throws EmployeesException{
-		
+
+	/* Adding a employees */
+	public Employees addingEmployee(Employees employee) throws EmployeesException {
+
 		try {
-			return employeesrepo.save(employee);
+
+			if (employee.getSalary() != 0.0) {
+				throw new EmployeesException("Salary is computed by payroll processing and should not be set manually. Please leave the salary field as null or zero.");
+			}
 			
+			return employeesrepo.save(employee);
+
 		} catch (Exception exception) {
 			throw new EmployeesException("Failed to add employee!");
 		}
-		
+
 	}
-	
-	
-	/*All employee*/
-	public List<Employees> findAllEmployee(){
+
+	/* All employee */
+	public List<Employees> findAllEmployee() {
 		return employeesrepo.findAll();
-		
+
 	}
-	
-	/*Getting employee by id*/
+
+	/* Getting employee by id */
 	public Employees findEmployeeById(int id) throws EmployeesException {
-		 return employeesrepo.findById(id).orElseThrow(() -> new EmployeesException(id + " : These id employee not found!"));
+		return employeesrepo.findById(id)
+				.orElseThrow(() -> new EmployeesException(id + " : These id employee not found!"));
 	}
-	
-	
-	/*Updating a employee*/
+
+	/* Updating a employee */
 	public Employees updateEmployee(int id, Employees employee) throws EmployeesException {
 		Optional<Employees> existingEmployee = employeesrepo.findById(id);
-		
-		if(existingEmployee.isPresent()) {
+
+		if (employee.getSalary() != 0.0) {
+			throw new EmployeesException(
+					"Salary is computed by payroll processing and should not be set manually or should be set to zero.");
+		}
+
+		if (existingEmployee.isPresent()) {
 			Employees updateEmployee = existingEmployee.get();
-			
+
 			updateEmployee.setName(employee.getName());
 			updateEmployee.setAge(employee.getAge());
 			updateEmployee.setGender(employee.getGender());
 			updateEmployee.setDesignation(employee.getDesignation());
 			updateEmployee.setRating(employee.getRating());
-			
+
 			return employeesrepo.save(updateEmployee);
-			
-		}else {
-			throw new EmployeesException(id +" : This employee id not found");
+
+		} else {
+			throw new EmployeesException(id + " : This employee id not found");
 		}
-		
+
 	}
-	
-	
-	/*Deleting a employee*/
-	public void deleteEmployee(int id) {
+
+	/* Deleting a employee */
+	public void deleteEmployee(int id) throws EmployeesException {
+		if (!employeesrepo.existsById(id)) {
+			throw new EmployeesException(id + " : This employee ID not found!");
+		}
 		employeesrepo.deleteById(id);
 	}
-	
-	
 
 }
